refactor: 경로 정리
