+++ src/app/layout.tsx
@@ -1,30 +1,23 @@
-import 'react-toastify/dist/ReactToastify.css';
-import './globals.css';
-
 import { Noto_Sans_KR } from 'next/font/google';
+import 'react-toastify/dist/ReactToastify.css';
 import { ToastContainer } from 'react-toastify';
 import * as sentry from '@sentry/nextjs';
 import type { Metadata } from 'next';
 import { ReactNode } from 'react';
+import './globals.css';
 import { QueryProvider } from '@/components';
 
 export const metadata: Metadata = {
   title: 'Velog Dashboard',
   description: 'Velog 통계를 확인할 수 있는 Velog Dashboard',
-  icons: {
-    icon: '/favicon.png',
-  },
+  icons: { icon: '/favicon.png' },
 };
 
-const NotoSansKr = Noto_Sans_KR({
-  subsets: ['latin'],
-});
+const NotoSansKr = Noto_Sans_KR({ subsets: ['latin'] });
 
 export default function RootLayout({
   children,
-}: Readonly<{
-  children: ReactNode;
-}>) {
+}: Readonly<{ children: ReactNode }>) {
   return (
     <html lang="ko">
       <body className={`${NotoSansKr.className} w-full h-screen bg-bg-main`}>
